chore(azure): working version executing checks
