+++ prowler/lib/check/check.py
@@ -526,6 +526,7 @@ def execute_checks(
                         f"Check '{check_name}' was not found for the {global_provider.provider.upper()} provider"
                     )
                 except Exception as error:
+                    print(traceback.print_last())
                     logger.error(
                         f"{check_name} - {error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
                     )
