+++ prowler/providers/azure/services/app/app_client.py
@@ -1,4 +1,4 @@
-from prowler.providers.azure.lib.audit_info.audit_info import azure_audit_info
 from prowler.providers.azure.services.app.app_service import App
+from prowler.providers.common.common import get_global_provider
 
-app_client = App(azure_audit_info)
+app_client = App(get_global_provider())
