+++ prowler/providers/azure/services/mysql/mysql_client.py
@@ -1,4 +1,4 @@
-from prowler.providers.azure.lib.audit_info.audit_info import azure_audit_info
 from prowler.providers.azure.services.mysql.mysql_service import MySQL
+from prowler.providers.common.common import get_global_provider
 
-mysql_client = MySQL(azure_audit_info)
+mysql_client = MySQL(get_global_provider())
