+++ prowler/providers/azure/services/policy/policy_client.py
@@ -1,4 +1,4 @@
-from prowler.providers.azure.lib.audit_info.audit_info import azure_audit_info
 from prowler.providers.azure.services.policy.policy_service import Policy
+from prowler.providers.common.common import get_global_provider
 
-policy_client = Policy(azure_audit_info)
+policy_client = Policy(get_global_provider())
