+++ prowler/providers/azure/services/vm/vm_client.py
@@ -1,4 +1,4 @@
-from prowler.providers.azure.lib.audit_info.audit_info import azure_audit_info
 from prowler.providers.azure.services.vm.vm_service import VirtualMachines
+from prowler.providers.common.common import get_global_provider
 
-vm_client = VirtualMachines(azure_audit_info)
+vm_client = VirtualMachines(get_global_provider())
