+++ prowler/providers/common/audit_info.py
@@ -4,13 +4,6 @@ from colorama import Fore, Style
 
 from prowler.config.config import load_and_validate_config_file
 from prowler.lib.logger import logger
-from prowler.providers.azure.azure_provider import Azure_Provider
-from prowler.providers.azure.lib.audit_info.audit_info import azure_audit_info
-from prowler.providers.azure.lib.audit_info.models import (
-    Azure_Audit_Info,
-    AzureRegionConfig,
-)
-from prowler.providers.azure.lib.exception.exception import AzureException
 from prowler.providers.gcp.gcp_provider import GCP_Provider
 from prowler.providers.gcp.lib.audit_info.audit_info import gcp_audit_info
 from prowler.providers.gcp.lib.audit_info.models import GCP_Audit_Info
@@ -48,19 +41,19 @@ Kubernetes Cluster: {Fore.YELLOW}[{cluster_name}]{Style.RESET_ALL}  User: {Fore.
 """
         print(report)
 
-    def print_azure_credentials(self, audit_info: Azure_Audit_Info):
-        printed_subscriptions = []
-        for key, value in audit_info.identity.subscriptions.items():
-            intermediate = f"{key} : {value}"
-            printed_subscriptions.append(intermediate)
-        report = f"""
-This report is being generated using the identity below:
+    #     def print_azure_credentials(self, audit_info: Azure_Audit_Info):
+    #         printed_subscriptions = []
+    #         for key, value in audit_info.identity.subscriptions.items():
+    #             intermediate = f"{key} : {value}"
+    #             printed_subscriptions.append(intermediate)
+    #         report = f"""
+    # This report is being generated using the identity below:
 
-Azure Tenant IDs: {Fore.YELLOW}[{" ".join(audit_info.identity.tenant_ids)}]{Style.RESET_ALL} Azure Tenant Domain: {Fore.YELLOW}[{audit_info.identity.domain}]{Style.RESET_ALL} Azure Region: {Fore.YELLOW}[{audit_info.azure_region_config.name}]{Style.RESET_ALL}
-Azure Subscriptions: {Fore.YELLOW}{printed_subscriptions}{Style.RESET_ALL}
-Azure Identity Type: {Fore.YELLOW}[{audit_info.identity.identity_type}]{Style.RESET_ALL} Azure Identity ID: {Fore.YELLOW}[{audit_info.identity.identity_id}]{Style.RESET_ALL}
-"""
-        print(report)
+    # Azure Tenant IDs: {Fore.YELLOW}[{" ".join(audit_info.identity.tenant_ids)}]{Style.RESET_ALL} Azure Tenant Domain: {Fore.YELLOW}[{audit_info.identity.domain}]{Style.RESET_ALL} Azure Region: {Fore.YELLOW}[{audit_info.azure_region_config.name}]{Style.RESET_ALL}
+    # Azure Subscriptions: {Fore.YELLOW}{printed_subscriptions}{Style.RESET_ALL}
+    # Azure Identity Type: {Fore.YELLOW}[{audit_info.identity.identity_type}]{Style.RESET_ALL} Azure Identity ID: {Fore.YELLOW}[{audit_info.identity.identity_id}]{Style.RESET_ALL}
+    # """
+    #         print(report)
 
     # TODO: remove if not needed, but not now :)
     # def set_aws_audit_info(self, arguments) -> AWS_Audit_Info:
@@ -303,64 +296,64 @@ Azure Identity Type: {Fore.YELLOW}[{audit_info.identity.identity_type}]{Style.RE
     #             f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
     #         )
     #         sys.exit(1)
+    # TODO: remove if not needed, but not now :)
+    # def set_azure_audit_info(self, arguments) -> Azure_Audit_Info:
+    #     """
+    #     set_azure_audit_info returns the Azure_Audit_Info
+    #     """
+    #     logger.info("Setting Azure session ...")
+    #     subscription_ids = arguments.get("subscription_ids")
 
-    def set_azure_audit_info(self, arguments) -> Azure_Audit_Info:
-        """
-        set_azure_audit_info returns the Azure_Audit_Info
-        """
-        logger.info("Setting Azure session ...")
-        subscription_ids = arguments.get("subscription_ids")
+    #     logger.info("Checking if any credentials mode is set ...")
+    #     az_cli_auth = arguments.get("az_cli_auth")
+    #     sp_env_auth = arguments.get("sp_env_auth")
+    #     browser_auth = arguments.get("browser_auth")
+    #     managed_entity_auth = arguments.get("managed_entity_auth")
+    #     tenant_id = arguments.get("tenant_id")
 
-        logger.info("Checking if any credentials mode is set ...")
-        az_cli_auth = arguments.get("az_cli_auth")
-        sp_env_auth = arguments.get("sp_env_auth")
-        browser_auth = arguments.get("browser_auth")
-        managed_entity_auth = arguments.get("managed_entity_auth")
-        tenant_id = arguments.get("tenant_id")
+    #     logger.info("Checking if region is different than default one")
+    #     region = arguments.get("azure_region")
 
-        logger.info("Checking if region is different than default one")
-        region = arguments.get("azure_region")
+    #     if (
+    #         not az_cli_auth
+    #         and not sp_env_auth
+    #         and not browser_auth
+    #         and not managed_entity_auth
+    #     ):
+    #         raise AzureException(
+    #             "Azure provider requires at least one authentication method set: [--az-cli-auth | --sp-env-auth | --browser-auth | --managed-identity-auth]"
+    #         )
+    #     if (not browser_auth and tenant_id) or (browser_auth and not tenant_id):
+    #         raise AzureException(
+    #             "Azure Tenant ID (--tenant-id) is required only for browser authentication mode"
+    #         )
 
-        if (
-            not az_cli_auth
-            and not sp_env_auth
-            and not browser_auth
-            and not managed_entity_auth
-        ):
-            raise AzureException(
-                "Azure provider requires at least one authentication method set: [--az-cli-auth | --sp-env-auth | --browser-auth | --managed-identity-auth]"
-            )
-        if (not browser_auth and tenant_id) or (browser_auth and not tenant_id):
-            raise AzureException(
-                "Azure Tenant ID (--tenant-id) is required only for browser authentication mode"
-            )
+    #     azure_provider = Azure_Provider(
+    #         az_cli_auth,
+    #         sp_env_auth,
+    #         browser_auth,
+    #         managed_entity_auth,
+    #         subscription_ids,
+    #         tenant_id,
+    #         region,
+    #     )
+    #     azure_audit_info.credentials = azure_provider.get_credentials()
+    #     azure_audit_info.identity = azure_provider.get_identity()
+    #     region_config = azure_provider.get_region_config()
+    #     azure_audit_info.azure_region_config = AzureRegionConfig(
+    #         name=region,
+    #         authority=region_config["authority"],
+    #         base_url=region_config["base_url"],
+    #         credential_scopes=region_config["credential_scopes"],
+    #     )
+    #     azure_audit_info.locations = azure_provider.get_locations(
+    #         azure_audit_info.credentials, region_config
+    #     )
 
-        azure_provider = Azure_Provider(
-            az_cli_auth,
-            sp_env_auth,
-            browser_auth,
-            managed_entity_auth,
-            subscription_ids,
-            tenant_id,
-            region,
-        )
-        azure_audit_info.credentials = azure_provider.get_credentials()
-        azure_audit_info.identity = azure_provider.get_identity()
-        region_config = azure_provider.get_region_config()
-        azure_audit_info.azure_region_config = AzureRegionConfig(
-            name=region,
-            authority=region_config["authority"],
-            base_url=region_config["base_url"],
-            credential_scopes=region_config["credential_scopes"],
-        )
-        azure_audit_info.locations = azure_provider.get_locations(
-            azure_audit_info.credentials, region_config
-        )
+    #     if not arguments.get("only_logs"):
+    #         self.print_azure_credentials(get_global_provider())
 
-        if not arguments.get("only_logs"):
-            self.print_azure_credentials(azure_audit_info)
-
-        return azure_audit_info
+    #     return azure_audit_info
 
     def set_gcp_audit_info(self, arguments) -> GCP_Audit_Info:
         """
@@ -417,6 +410,7 @@ def set_provider_audit_info(provider: str, arguments: dict):
         provider_audit_info = getattr(Audit_Info(), provider_set_audit_info)(arguments)
 
         # Set the audit configuration from the config file
+        # TODO: move this to the providers
         provider_audit_info.audit_config = load_and_validate_config_file(
             provider, arguments["config_file"]
         )
