+++ prowler/providers/common/provider.py
@@ -12,3 +12,10 @@ class Provider(ABC):
 
     def validate_arguments(self):
         pass
+
+    def get_checks_to_execute_by_audit_resources(self):
+        """
+        get_checks_to_execute_by_audit_resources returns a set of checks based on the input resources to scan.
+
+        This is a fallback that returns None if the service has not implemented this function.
+        """
